# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SMS Banking Bridge API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface RegisterUserRequest {
  phoneNumber: string;
  firstName: string;
  lastName: string;
  smsConsent: boolean;
}

export type UserOnboardingStatus =
  (typeof UserOnboardingStatus)[keyof typeof UserOnboardingStatus];

export const UserOnboardingStatus = {
  pending: "pending",
  bank_linked: "bank_linked",
  active: "active",
  opted_out: "opted_out",
} as const;

export interface User {
  id: number;
  phoneNumber: string;
  firstName: string;
  lastName: string;
  smsConsent: boolean;
  consentDate?: string;
  optedOut: boolean;
  onboardingStatus: UserOnboardingStatus;
  createdAt: string;
}

export interface LinkedAccount {
  id: number;
  userId: number;
  bankName: string;
  accountType: string;
  accountLastFour: string;
  nickname: string;
  currentBalance: number;
  createdAt: string;
}

export type UserWithAccounts = User & {
  accounts?: LinkedAccount[];
  lastSmsAt?: string | null;
};

export type LinkAccountRequestAccountType =
  (typeof LinkAccountRequestAccountType)[keyof typeof LinkAccountRequestAccountType];

export const LinkAccountRequestAccountType = {
  checking: "checking",
  savings: "savings",
  credit: "credit",
} as const;

export interface LinkAccountRequest {
  userId: number;
  bankName: string;
  accountType: LinkAccountRequestAccountType;
  accountLastFour: string;
  nickname: string;
  currentBalance: number;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  debit: "debit",
  credit: "credit",
} as const;

export interface Transaction {
  id: number;
  accountId: number;
  userId: number;
  description: string;
  amount: number;
  type: TransactionType;
  category: string;
  merchantName: string;
  transactionDate: string;
  runningBalance: number;
}

export type SmsLogDirection =
  (typeof SmsLogDirection)[keyof typeof SmsLogDirection];

export const SmsLogDirection = {
  inbound: "inbound",
  outbound: "outbound",
} as const;

export type SmsLogStatus = (typeof SmsLogStatus)[keyof typeof SmsLogStatus];

export const SmsLogStatus = {
  delivered: "delivered",
  pending: "pending",
  failed: "failed",
} as const;

export interface SmsLog {
  id: number;
  userId: number;
  phoneNumber: string;
  direction: SmsLogDirection;
  message: string;
  command?: string | null;
  status: SmsLogStatus;
  createdAt: string;
}

export interface SimulateSmsRequest {
  userId: number;
  command: string;
}

export interface SimulateSmsResponse {
  command: string;
  response: string;
  logId: number;
}

export interface AdminStats {
  totalUsers: number;
  activeUsers: number;
  optedOutUsers: number;
  totalSmsToday: number;
  totalAccountsLinked: number;
  pendingOnboarding: number;
}

export interface AlertSettings {
  alertsEnabled: boolean;
  weekendPauseEnabled: boolean;
  /** Day and time to start pause (e.g. friday_14:00) */
  weekendPauseStart: string;
  /** Day and time to end pause (e.g. sunday_00:00) */
  weekendPauseEnd: string;
  maxDailyAlerts: number;
  lowBalanceThreshold: number;
  largeTransactionThreshold: number;
}

export type GetUserTransactionsParams = {
  accountId?: number;
  limit?: number;
};

export type GetSmsLogsParams = {
  userId?: number;
  limit?: number;
};