# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SMS Banking Bridge API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const RegisterUserBody = zod.object({
  phoneNumber: zod.string(),
  firstName: zod.string(),
  lastName: zod.string(),
  smsConsent: zod.boolean(),
});

/**
 * @summary List all users (admin)
 */
export const ListUsersResponseItem = zod
  .object({
    id: zod.number(),
    phoneNumber: zod.string(),
    firstName: zod.string(),
    lastName: zod.string(),
    smsConsent: zod.boolean(),
    consentDate: zod.date().optional(),
    optedOut: zod.boolean(),
    onboardingStatus: zod.enum([
      "pending",
      "bank_linked",
      "active",
      "opted_out",
    ]),
    createdAt: zod.date(),
  })
  .and(
    zod.object({
      accounts: zod
        .array(
          zod.object({
            id: zod.number(),
            userId: zod.number(),
            bankName: zod.string(),
            accountType: zod.string(),
            accountLastFour: zod.string(),
            nickname: zod.string(),
            currentBalance: zod.number(),
            createdAt: zod.date(),
          }),
        )
        .optional(),
      lastSmsAt: zod.date().nullish(),
    }),
  );
export const ListUsersResponse = zod.array(ListUsersResponseItem);

/**
 * @summary Get a user by ID
 */
export const GetUserParams = zod.object({
  userId: zod.coerce.number(),
});

export const GetUserResponse = zod
  .object({
    id: zod.number(),
    phoneNumber: zod.string(),
    firstName: zod.string(),
    lastName: zod.string(),
    smsConsent: zod.boolean(),
    consentDate: zod.date().optional(),
    optedOut: zod.boolean(),
    onboardingStatus: zod.enum([
      "pending",
      "bank_linked",
      "active",
      "opted_out",
    ]),
    createdAt: zod.date(),
  })
  .and(
    zod.object({
      accounts: zod
        .array(
          zod.object({
            id: zod.number(),
            userId: zod.number(),
            bankName: zod.string(),
            accountType: zod.string(),
            accountLastFour: zod.string(),
            nickname: zod.string(),
            currentBalance: zod.number(),
            createdAt: zod.date(),
          }),
        )
        .optional(),
      lastSmsAt: zod.date().nullish(),
    }),
  );

/**
 * @summary Link a bank account to a user
 */
export const LinkAccountBody = zod.object({
  userId: zod.number(),
  bankName: zod.string(),
  accountType: zod.enum(["checking", "savings", "credit"]),
  accountLastFour: zod.string(),
  nickname: zod.string(),
  currentBalance: zod.number(),
});

/**
 * @summary Get accounts for a user
 */
export const GetUserAccountsParams = zod.object({
  userId: zod.coerce.number(),
});

export const GetUserAccountsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  bankName: zod.string(),
  accountType: zod.string(),
  accountLastFour: zod.string(),
  nickname: zod.string(),
  currentBalance: zod.number(),
  createdAt: zod.date(),
});
export const GetUserAccountsResponse = zod.array(GetUserAccountsResponseItem);

/**
 * @summary Get recent transactions for a user
 */
export const GetUserTransactionsParams = zod.object({
  userId: zod.coerce.number(),
});

export const getUserTransactionsQueryLimitDefault = 10;

export const GetUserTransactionsQueryParams = zod.object({
  accountId: zod.coerce.number().optional(),
  limit: zod.coerce.number().default(getUserTransactionsQueryLimitDefault),
});

export const GetUserTransactionsResponseItem = zod.object({
  id: zod.number(),
  accountId: zod.number(),
  userId: zod.number(),
  description: zod.string(),
  amount: zod.number(),
  type: zod.enum(["debit", "credit"]),
  category: zod.string(),
  merchantName: zod.string(),
  transactionDate: zod.date(),
  runningBalance: zod.number(),
});
export const GetUserTransactionsResponse = zod.array(
  GetUserTransactionsResponseItem,
);

/**
 * @summary Get SMS activity logs (admin)
 */
export const getSmsLogsQueryLimitDefault = 50;

export const GetSmsLogsQueryParams = zod.object({
  userId: zod.coerce.number().optional(),
  limit: zod.coerce.number().default(getSmsLogsQueryLimitDefault),
});

export const GetSmsLogsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  phoneNumber: zod.string(),
  direction: zod.enum(["inbound", "outbound"]),
  message: zod.string(),
  command: zod.string().nullish(),
  status: zod.enum(["delivered", "pending", "failed"]),
  createdAt: zod.date(),
});
export const GetSmsLogsResponse = zod.array(GetSmsLogsResponseItem);

/**
 * @summary Simulate an SMS command
 */
export const SimulateSmsBody = zod.object({
  userId: zod.number(),
  command: zod.string(),
});

export const SimulateSmsResponse = zod.object({
  command: zod.string(),
  response: zod.string(),
  logId: zod.number(),
});

/**
 * @summary Get admin dashboard stats
 */
export const GetAdminStatsResponse = zod.object({
  totalUsers: zod.number(),
  activeUsers: zod.number(),
  optedOutUsers: zod.number(),
  totalSmsToday: zod.number(),
  totalAccountsLinked: zod.number(),
  pendingOnboarding: zod.number(),
});

/**
 * @summary Get global alert settings
 */
export const GetAlertSettingsResponse = zod.object({
  alertsEnabled: zod.boolean(),
  weekendPauseEnabled: zod.boolean(),
  weekendPauseStart: zod
    .string()
    .describe("Day and time to start pause (e.g. friday_14:00)"),
  weekendPauseEnd: zod
    .string()
    .describe("Day and time to end pause (e.g. sunday_00:00)"),
  maxDailyAlerts: zod.number(),
  lowBalanceThreshold: zod.number(),
  largeTransactionThreshold: zod.number(),
});

/**
 * @summary Update global alert settings
 */
export const UpdateAlertSettingsBody = zod.object({
  alertsEnabled: zod.boolean(),
  weekendPauseEnabled: zod.boolean(),
  weekendPauseStart: zod
    .string()
    .describe("Day and time to start pause (e.g. friday_14:00)"),
  weekendPauseEnd: zod
    .string()
    .describe("Day and time to end pause (e.g. sunday_00:00)"),
  maxDailyAlerts: zod.number(),
  lowBalanceThreshold: zod.number(),
  largeTransactionThreshold: zod.number(),
});

export const UpdateAlertSettingsResponse = zod.object({
  alertsEnabled: zod.boolean(),
  weekendPauseEnabled: zod.boolean(),
  weekendPauseStart: zod
    .string()
    .describe("Day and time to start pause (e.g. friday_14:00)"),
  weekendPauseEnd: zod
    .string()
    .describe("Day and time to end pause (e.g. sunday_00:00)"),
  maxDailyAlerts: zod.number(),
  lowBalanceThreshold: zod.number(),
  largeTransactionThreshold: zod.number(),
});