# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SMS Banking Bridge API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalUsers: number;
  activeUsers: number;
  optedOutUsers: number;
  totalSmsToday: number;
  totalAccountsLinked: number;
  pendingOnboarding: number;
}