# lib/api-zod/src/generated/types/linkAccountRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SMS Banking Bridge API
 * OpenAPI spec version: 0.1.0
 */
import type { LinkAccountRequestAccountType } from "./linkAccountRequestAccountType";

export interface LinkAccountRequest {
  userId: number;
  bankName: string;
  accountType: LinkAccountRequestAccountType;
  accountLastFour: string;
  nickname: string;
  currentBalance: number;
}