# lib/api-zod/src/generated/types/simulateSmsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SMS Banking Bridge API
 * OpenAPI spec version: 0.1.0
 */

export interface SimulateSmsResponse {
  command: string;
  response: string;
  logId: number;
}