# lib/api-zod/src/generated/types/smsLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SMS Banking Bridge API
 * OpenAPI spec version: 0.1.0
 */
import type { SmsLogDirection } from "./smsLogDirection";
import type { SmsLogStatus } from "./smsLogStatus";

export interface SmsLog {
  id: number;
  userId: number;
  phoneNumber: string;
  direction: SmsLogDirection;
  message: string;
  command?: string | null;
  status: SmsLogStatus;
  createdAt: Date;
}