# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SMS Banking Bridge API
 * OpenAPI spec version: 0.1.0
 */
import type { UserOnboardingStatus } from "./userOnboardingStatus";

export interface User {
  id: number;
  phoneNumber: string;
  firstName: string;
  lastName: string;
  smsConsent: boolean;
  consentDate?: Date;
  optedOut: boolean;
  onboardingStatus: UserOnboardingStatus;
  createdAt: Date;
}