# lib/api-zod/src/generated/types/userWithAccounts.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SMS Banking Bridge API
 * OpenAPI spec version: 0.1.0
 */
import type { LinkedAccount } from "./linkedAccount";
import type { User } from "./user";

export type UserWithAccounts = User & {
  accounts?: LinkedAccount[];
  lastSmsAt?: Date | null;
};